Add implicit API wrappers for DriveFile and DriveFolder
Add @delegateApi annotations that works just like @provideApi, but delegates the calls not to a static object, but to a implicit class parameter. This is useful for APIs which return types that themselves provide API-methods taking a GoogleApiClient or return PendingResults.
First use-cases of that annotation are DriveFile and DriveFolder wrappers in the Drive API.